LT-20432 (follow-on): Fix Data Deletion
* Hold only the ID's of WS's to be deleted.
Holding the entire definition prevents most data
from being deleted (don't ask me why).

+++ Src/FwCoreDlgs/FwWritingSystemSetupModel.cs
@@ -54,7 +54,7 @@ namespace SIL.FieldWorks.FwCoreDlgs
 		private readonly IWritingSystemManager _wsManager;
 		private string _languageName;
 		private WritingSystemSetupModel _currentWsSetupModel;
-		private readonly ISet<CoreWritingSystemDefinition> _wssToDelete = new HashSet<CoreWritingSystemDefinition>();
+		private readonly ISet<string> _wsIdsToDelete = new HashSet<string>();
 		private readonly Dictionary<CoreWritingSystemDefinition, CoreWritingSystemDefinition> _mergedWritingSystems = new Dictionary<CoreWritingSystemDefinition, CoreWritingSystemDefinition>();
 
 
@@ -807,7 +807,7 @@ namespace SIL.FieldWorks.FwCoreDlgs
 		{
 			var removedWritingSystems = new List<CoreWritingSystemDefinition>(allWritingSystems);
 			removedWritingSystems.RemoveAll(ws => workingWritingSystems.Any(wws => wws.Id == ws.Id));
-			_wssToDelete.RemoveAll(ws => workingWritingSystems.Any(wws => wws.Id == ws.Id));
+			_wsIdsToDelete.RemoveAll(wsId => workingWritingSystems.Any(wws => wws.Id == wsId));
 			foreach (var deleteCandidate in removedWritingSystems)
 			{
 				currentWritingSystems.Remove(deleteCandidate);
@@ -821,13 +821,14 @@ namespace SIL.FieldWorks.FwCoreDlgs
 			}
 
 			var deletedWsIds = new List<string>();
-			foreach (var deleteCandidate in _wssToDelete)
+			foreach (var deleteCandidateId in _wsIdsToDelete)
 			{
-				if (!otherWritingSystems.Contains(deleteCandidate)
+				if (Cache.ServiceLocator.WritingSystemManager.TryGet(deleteCandidateId, out var deleteCandidate)
+					&& !otherWritingSystems.Contains(deleteCandidate)
 					&& !_mergedWritingSystems.Keys.Contains(deleteCandidate))
 				{
 					WritingSystemServices.DeleteWritingSystem(Cache, deleteCandidate);
-					deletedWsIds.Add(deleteCandidate.Id);
+					deletedWsIds.Add(deleteCandidateId);
 				}
 			}
 
@@ -916,7 +917,7 @@ namespace SIL.FieldWorks.FwCoreDlgs
 
 		private void HideCurrentWritingSystem(object sender, EventArgs e)
 		{
-			_wssToDelete.Remove(_wssToDelete.FirstOrDefault(ws => ws.Id == _currentWs.Id));
+			_wsIdsToDelete.Remove(_wsIdsToDelete.FirstOrDefault(wsId => wsId == _currentWs.Id));
 			HideCurrentWritingSystem();
 		}
 
@@ -934,7 +935,7 @@ namespace SIL.FieldWorks.FwCoreDlgs
 
 			if (ConfirmDeleteWritingSystem(CurrentWsSetupModel.CurrentDisplayLabel)) // prompt the user to delete the WS and its data
 			{
-				_wssToDelete.Add(_currentWs);
+				_wsIdsToDelete.Add(_currentWs.Id);
 				HideCurrentWritingSystem();
 			}
 		}
@@ -1093,11 +1094,11 @@ namespace SIL.FieldWorks.FwCoreDlgs
 		private void ViewHiddenWritingSystemsHandler(object sender, EventArgs e)
 		{
 			var model = new ViewHiddenWritingSystemsModel(_listType, Cache,
-					WorkingList.Select(li => li.OriginalWs).Where(ws => ws != null).ToList(), _wssToDelete)
+					WorkingList.Select(li => li.OriginalWs).Where(ws => ws != null).ToList(), _wsIdsToDelete)
 				{ ConfirmDeleteWritingSystem = ConfirmDeleteWritingSystem };
 			ViewHiddenWritingSystems(model);
 
-			_wssToDelete.AddRange(model.DeletedWritingSystems);
+			_wsIdsToDelete.AddRange(model.DeletedWritingSystems.Select(ws => ws.Id));
 			foreach (var addedWS in model.AddedWritingSystems)
 			{
 				AddNewLanguage(new LanguageInfo{DesiredName = addedWS.LanguageName, LanguageTag = addedWS.LanguageTag});
